chore(integrations): rename utils.ts to const.ts
